Reconfigure database with `californica` values (#37)
* Reconfigure database with `californica` values

Add database configuration for `postgresql`; update `.env.sample` to include
all currently configurable values. Lastly, add `dotenv` documentation to
README.md.

* Passwordless and secure PostgreSQL howto
added info on how to enable passwordless authentication for PostgreSQL
added info on how to define a password for a role
added info on how to use a role with a password with .env files

* Revise README to point to new dev env wiki

Dropped most of the setup guidelines from the README, replaced with
links to the development environment pages on the Amalgamated Samvera
wiki. Also changed the last remaining mentions of Tenejo in the README,
because they're bothersome.