Show metadata terms on show page
This adds additonal setup for the existing custom
metadata terms so that they can be displayed on the
show view of an Work.

In order for them to be used by the `WorkPresenter`
they need to explicilty indexed in the model.

The `WorkPresenter` needs to have references to the
to the terms.

The `_attribute_rows.html.erb` view needs to be
overidden to include the additional attributes.

The `catalog_controller` needs to be updated with
all attributes that you want to see in the view.

There is a view spec that tests for the presence of the
the attributes on that page.

Connected to #195