Add error_stream and info_stream parameters to our custom parser
We need to log errors and info directly from the parser to give us the ability
to track total records processed (could we implement a progress-reporting info
logger instead?). To allow for this without hard coding a direct reach into the
global config, we add parameters to our custom parser.