Logging refactor (#217)
* Introduce a `CalifornicaLogStream` and use it in the Importer

This reduces the footprint of our log streams to match the darlingtonia stream
definition `#<<`. We want to be able to interoperate with (or simply switch to)
streams used by other Darlingtonia projects. This will eventually replace the
`MessageStream` class.

* Deduplicate logging/error handling behavior in `CalifornicaCsvParser`

`Darlingtonia::RecordImporter` is the appropriate place logging and error
handling behavior for individual record failures. This behavior is included by
default in the base implementation:
https://github.com/curationexperts/darlingtonia/blob/b7432e2c471c904ee86145dab0b4189d4fbd9587/lib/darlingtonia/record_importer.rb#L24-L31

The behavior was largely duplicated in `CalifornicaCsvParser`, and can be simply
removed.

* Log import errors at ERROR level

Passing the error logger directly to the `RecordImporter` ensures we use our
custom streams, instead the default ones.

* Add error_stream and info_stream parameters to our custom parser

We need to log errors and info directly from the parser to give us the ability
to track total records processed (could we implement a progress-reporting info
logger instead?). To allow for this without hard coding a direct reach into the
global config, we add parameters to our custom parser.

* Remove global side-effect in `Californica::Importer#initialize`

Initializing a `Californica::Importer` preivously set `Darlingtonia.config` as a
global side-effect. Following some refactors to change the error and info stream
setting behavior, this is no longer neccessary to log to the expected places.

Removing the side effect makes code less fragile and fixes some order-reliant
tests.

* Remove dead code in `MessageStream`

This has been replaced by `CalifornicaLogStream`. All previous callers have been
removed.