Ingest log (#252)
* Create a list of missing files during ingest

Refactor flaky test
Allow uploads from fixtures directory
Set sensible default for missing files log
Also log when an item is missing a file entirely

* Cleanup rake tasks

Clean separation between delete all and
a full ingest. We should be able to do
a full ingest without waiting for a delete
if we're already in an empty environment.