Customize the ingest actor stack
When we're ingesting content, only call the parts
of the actor stack that we really need for creating
ingested objects. However, keep the default actor
stack untouched, so that we don't break the ability
to create and edit objects via the UI.

Add rake tasks for profiling and benchmarking
the ingest of a single record.

This seems to significantly improve performance:

Before removal:
------
Benchmark for ingest of 21 sample records (elapsed real time is last):
 14.400951   0.885906  15.286857 ( 32.241227)

After removal:
------
Benchmark for ingest of 21 sample records (elapsed real time is last):
  7.453626   0.650355   8.103981 ( 15.822791)