Implement a read-only mode (#488)
As someone who cares about data preservation, I want a read-only mode for the system. When californica is in read-only mode it can still be searched and ursus will will be able to read from californica, but no changes can be made. This will create a safe window during which backups can be performed.

 There is a feature flipper that will turn read only mode on and off
 When read-only mode is in effect there is a banner telling me why I can't add or edit records