If the work has a parent ARK in the CSV, add the work to that collection. Story #432.
* The importer now supports a new column in the CSV file called "Parent
ARK".  If a work has a Parent ARK, that ARK will be used to find the
correct Collection to add the Work record to.  If the Collection doesn't
exist yet, a new Collection will be created using that ARK.

* The ARKs in the input CSV files don't usually have the "ark:/" prefix
on the strings.  We had previously written some code in the indexer
classes that adds the prefix onto the strings.  We need to use that same
logic during the import, so I pulled that code into a separate Ark
class.

* I added member_of_collections_attributes method to the mapper class to
set up the proper attributes hash so that the actor stack will add the
work to the correct collection.

* I updated the find_or_create_by_ark method to create a permission
template for the collection (following the example from Hyrax, when a
collection gets created in
app/controllers/hyrax/dashboard/collections_controller.rb).  Without a
permission template, the actor stack will fail when you run the
importer.