Import a Collection record from the CSV. Story #436
* Added a CollectionRecordImporter class that will be used (instead of
the existing ActorRecordImporter) to create or update Collection
records.

* The new CollectionRecordImporter needed to use the same overridden
find_existing_record method that the ActorRecordImporter uses, so I
pulled that method into a module that is included in both classes.

* Added a RecordImporter class that will act as a wrapper to select and
use the correct type of record importer class for each record.

* In the CalifornicaMapper, I changed the rights_statement to be an
array, as it properly should be.  In the past we could be sloppy about
that because the actor stack will cast it to an array.