Reverse the ID value
This will give us an even b-tree distribution in fedora, which will
improve performance as the repository grows.

"ModeShape has been designed to efficiently handle a single node having a
large number (100K-200k) of child nodes out-of-the-box. However, if an
application stores a lot more nodes than that under a single parent,
performance in terms of latency and memory consumption starts to
decrease proportionally to the number of stored children."
Source:
https://docs.jboss.org/author/display/MODE40/Large+numbers+of+child+nodes#Largenumbersofchildnodes-Unorderedlargecollections

Since most (all?) of the UCLA ark identifiers start with the same
string, unless we reverse the ID values we'll have everything stored
under the same node in Fedora and hit performance issues when the system
reaches > 200k objects.

This PR also adds unhappy path tests to ARK-->ID generation