Use FITS Servlet
This commit allows you to use an external
FITS Servlet for file characterization. To turn
it on, set the 'FITS_SERVLET_URL' environment varible
on your server before starting the server. This is
accomplished with overrides and prepends, so it won't
be possible for a FlipFlop feature to toggle this, because
you will need to restart your server.

The FITS servlet URL should be the full path to the FITS endpoint.

In the `docker-compose` environment this is at `http://fits:8080/fits`.
In the `docker-compose` environment the specs are ran against the
remote fits at `fits:8080`. On travis, the default installation is used.

If the `FITS_SERVLET_URL` is not set, the server will use the default
FITS installation.

Connected to CAL-580