Added a new value for the visibility of a work. Story #608
* You can now set a value of "discovery" for the visibility of a Work or
FileSet.  The intention is that when we display records with
"discovery"-level access in Ursus, the user will be able to see the
record's metadata, but not the attached files or images.

* The hydra-access-controls and blacklight-access_controls gems already
have the concept of "discovery"-level access, but that level of
visibility seems to be disabled/unused in the hyrax gem.  This code
re-enables that existing visibility level for Californica records.  This
will allow us to use the blacklight-access_controls gem to enforce
proper visibility in Ursus.

* I added the Discoverable module to override the visibility from the
hydra-access-controls gem and allow us to add a new value.  This module
is included in the Work, ChildWork, and FileSet classes.

* I overrode the visibility method in the SolrDocument class to be able
to properly display the new value in the UI.

* Added specs for SolrDocument to make sure that the new visibility
doesn't interfere with Hyrax's lease and embargo behavior.

* I copied some form partials from Hyrax to add the new visibility into
the edit forms.
_form_visibility_component.html.erb and _form_permission.html.erb

* Out of scope:  Setting "discovery" access in a lease or embargo.  We
can add that functionality in the future if it is ever needed in
Californica.