Generate IIIF Manifest for works with ChildWorks
The current views and libraries that generate the IIIF manifest
and display the UV interface assume that the the
`Work` is structured with an attached `FileSet` not additional
works. The first `Work` requires an attached `FileSet` or
it will not display.

The overrides the generation of the manifest so that
first `Work` will generate a manifest even if it  does
not have an image/FileSet attached to it.

This is done with a JBuilder view that the manifest
is genreated from.

Doing it this way it significantly faster. An Ethiopic manuscript
was taking 5 minutes to genrate a manifest. This is down to 1 minute.

That's still too slow, but it is still a large reduction.

Connected to CAL-638