Fixed a problem with access controls in a FileSet record. Part of story 608.
I noticed this problem in the UI: when you change the visibility of a
work record, it gives you the option to apply those changes to all the
child records.  If you select 'yes' background jobs run to update all
the attached FileSet records.  But in some situations the resulting
visibility for the FileSets didn't match the parent work.

The problem was caused by an assumption in the hydra-access-controls
gem that caused it to incorrectly match on some access control records
that it was selecting for removal.  Because we are overriding visibility
in Californica, that assumption is no longer true.  I'll take a closer
look at the hydra-access-controls gem later and decide if we should file
a bug against it, but for now, overriding the "permissions_attributes="
method from that gem allows us to fix the invalid assumption.

I added a spec for "inherit_permissions_job_spec", which tests the exact
problem at a higher level than a unit test would.