Fix - Edit page fails to render visibility badge in dev (not on servers)
Edit page failed to render visibility badge in dev (not on servers)

The work edit page failed to render due to an error related to the "discovery"-level visibility badge.

This happened only in development, not in production. That's because this line is invoked in production:
https://github.com/UCLALibrary/californica/blob/a6e251a1692aecb073a9928d6d6f1c91e9b38700/config/initializers/hyrax.rb#L13

But per this discussion:
https://stackoverflow.com/questions/8895103/how-can-i-keep-my-initializer-configuration-from-being-lost-in-development-mode
it ran once in development but then reset on subsequent requests.

---

Changes to be committed:
+ modified: `config/initializers/hyrax.rb`