techdebt: Combine examples in import_and_show_work_spec.rb
As written, the test was running the import for each 'it' block. While the best practice for unit tests is to give each assertion it's own 'it' block, for system (aka integration) test this gets very slow, and it makes more sense to just do the thing once then make a bunch of assertions about it.

This change reduces the runtime for this spec from 35 to 19 seconds.