derive CsvImport status by inspecting sub-jobs
Previously CsvImport status and elapsed_time were set at the end of StartCsvImportJob. This worked when the various sub-jobs were executed inline, but not that they are queued and executed in parallel it is not appropriate.

Instead, staus and elapsed_time are set by CsvImportService.update_status, which executes at the end of each sub-job.

update_status queries all sub-tasks of a CsvImport (CsvRow and CsvImportTask objects). If all have statuses of 'complete' or 'error', then it:
- sets the CsvImport status to 'complete'.
- sets the CsvImport end_time to that of the last sub-task to complete, and calculates elapsed_time and elapsed_time_per_record accordingly.