When CSV Import restarts, import all rows that aren't 'complete' or 'pending finalization'
Previously, only 'queued' rows were imported. A previous PR added this behavior for 'in progress', but the new 'deleting child works' status was not restarted.

The new approach guarantees all such cases would be restarted, as well as that rows with errors will be retried (useful if manually restarting after fixing a bug).